requirements.txt
yfinance
pandas
psycopg2-binary
apache-airflow-providers-postgres
apache-airflow-providers-http
requests
python-dotenv

